refactor(tests): remove deprecated __tests__ folder for eslint_custom_rules
