+++ eslint-custom-rules/nbsp-in-text.test.js
@@ -1,7 +1,7 @@
 import { RuleTester } from 'eslint'
 import { config } from './config'
 
-import rule from '../nbsp-in-text'
+import rule from './nbsp-in-text'
 
 const ruleTester = new RuleTester()
 
@@ -27,13 +27,12 @@ const tests = {
   invalid: [
     { code: `'myText !'`, output: `'myText\\u00a0!'`, errors: 1 },
     { code: `'« myText'`, output: `'«\\u00a0myText'`, errors: 1 },
-    { code: `\`myText :\`` ,output: `\`myText\\u00a0:\``, errors: 1 },
-    { code: `\`« myText\`` ,output: `\`«\\u00a0myText\``, errors: 1 },
-    { code: `"myText ?"` ,output: `"myText&nbsp;?"`, errors: 1 },
-    { code: `"« myText"` ,output: `"«&nbsp;myText"`, errors: 1 },
-    { code: '<Text>myText €</Text>' ,output: '<Text>myText&nbsp;€</Text>', errors: 1 },
-    { code: '<Text>« myText</Text>' ,output: '<Text>«&nbsp;myText</Text>', errors: 1 },
-
+    { code: `\`myText :\``, output: `\`myText\\u00a0:\``, errors: 1 },
+    { code: `\`« myText\``, output: `\`«\\u00a0myText\``, errors: 1 },
+    { code: `"myText ?"`, output: `"myText&nbsp;?"`, errors: 1 },
+    { code: `"« myText"`, output: `"«&nbsp;myText"`, errors: 1 },
+    { code: '<Text>myText €</Text>', output: '<Text>myText&nbsp;€</Text>', errors: 1 },
+    { code: '<Text>« myText</Text>', output: '<Text>«&nbsp;myText</Text>', errors: 1 },
   ],
 }
