+++ eslint-custom-rules/no-raw-text.test.js
@@ -1,7 +1,7 @@
 import { RuleTester } from 'eslint'
 import { config } from './config'
 
-import rule from '../no-raw-text'
+import rule from './no-raw-text'
 
 const ruleTester = new RuleTester()
