+++ eslint-custom-rules/no-use-of-algolia-multiple-queries.test.js
@@ -1,17 +1,15 @@
 import { RuleTester } from 'eslint'
 import { config } from './config'
 
-import rule from '../no-use-of-algolia-multiple-queries'
+import rule from './no-use-of-algolia-multiple-queries'
 
 const ruleTester = new RuleTester()
 
 const tests = {
-  valid: [
-    { code: "await multipleQueries(queries)" },
-  ],
+  valid: [{ code: 'await multipleQueries(queries)' }],
   invalid: [
     {
-      code: "await client.multipleQueries(queries)",
+      code: 'await client.multipleQueries(queries)',
       errors: 1,
     },
   ],
